src/Services/Detections/ai_inferencer/requirements.txt

# For Dapr ------------------------------------
dapr-ext-grpc
dapr
setuptools==58.2.0

# For Events serialization ------------------------------------
avro-python3
avro-json-serializer

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
# opencv-python-headless>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0

pandas>=1.1.4

# Plotting ------------------------------------
# CDLTLL # seaborn>=0.11.0

# Extras --------------------------------------
psutil  # system utilization
smmap
gitdb
gitpython
IPython

# For YOLO model ------------------------------------
# CDLTLL # ultralytics